requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
flask==3.0.0
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
flask-limiter==3.5.0

